split up data provider
